Fixes #4 - Conditionally generates state objects

+++ src/BaseModelGenerator.js
@@ -10,17 +10,18 @@ import Handlebars from 'handlebars'    // Template language for code generation
 var fs = require('fs')                 // Read/write files
 
 export default class BaseModelGenerator {
-  constructor(modelObject, templateFileName, outputFileName) {
+  constructor(modelObject, templateFileName, outputFileName, overwrite = true) {
     this._modelObject = modelObject            // A redux-immutable-object class
     this._name = modelObject._name
     this._templateFileName = templateFileName  // File name containing the template
     this._outputFileName = outputFileName      // File name for the generator output
+    this._overwrite = overwrite                // Flag if existing files should be overwritten
   }
 
   // This is a base method intended to be overridden in subclasses; it should return
   // an initialized context for the specific source file being generated
   getInitialContext() {
-    return { name: this._modelObject.name }
+    return { name: this._name }
   }
  
   // This is a base method intended to be overridden in subclasses; it should update
@@ -49,11 +50,14 @@ export default class BaseModelGenerator {
   // Implements the default behavior of reading the template file, creating the
   // context, and rendering the resulting source file
   render() {
-    const template = Handlebars.compile(fs.readFileSync(this._templateFileName, 'utf8'))
-    const myContext = this.getInitialContext()
-    this.populateContext(myContext)
-    this.finalizeContext(myContext)
-    const result = template(myContext)
-    fs.writeFileSync(this._outputFileName, result)
+    // Only render if overwrites are OK or target file does not exist
+    if (this._overwrite || !fs.existsSync(this._outputFileName)) {
+      const template = Handlebars.compile(fs.readFileSync(this._templateFileName, 'utf8'))
+      const myContext = this.getInitialContext()
+      this.populateContext(myContext)
+      this.finalizeContext(myContext)
+      const result = template(myContext)
+      fs.writeFileSync(this._outputFileName, result)
+    }
   }
 }
